Fix interpreter specification
Hard-coding the selected interpreter to be `#!/usr/bin/ruby -w` causes problems in environments where a system `ruby` is not installed (or is at least not installed *there*, as is the case on systems where `ruby` is actually in `/usr/local/bin`).  Specifying `#!/usr/bin/env ruby -w` is more flexible and generally good practice.